requirements.txt
google-api-python-client==2.65.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
tinydb==4.7.0
python-telegram-bot==13.14
SQLAlchemy==1.4.43
mysqlclient==2.1.1
speechrecognition==3.8.1
pydub==0.25.1